requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.4
beautifulsoup4==4.9.3
black==20.8b1
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.5
click==7.1.2
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2
django-bootstrap4==3.0.0
django-solo==1.1.5
future==0.18.2
html5lib==1.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
pytz==2021.1
regex==2021.4.4
reportlab==3.5.66
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
webencodings==0.5.1
xhtml2pdf==0.2.5
